requirements.txt
Flask
flask-cors
flask-jwt-extended
bcrypt
pymongo
python-dotenv
gunicorn
